moved sql and node info related functions for ethereum layer to separate files function names renamed bug fix
